Rename "rootsgetter" package to "roots" and shorten func name
Removes some stutter from func calls.
